+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - `--version` flag to display installed version
+- Auto-detection of `caption` package for `\caption*` (unnumbered captions) and `\captionof` commands
 
 ## [1.3.0] - 2025-12-09